refactor: CLI config
