+++ packages/daf-cli/src/cli.ts
@@ -14,6 +14,7 @@ import './crypto'
 import './execute'
 import './server'
 import './setup'
+import './config'
 import './schema'
 
 if (!process.argv.slice(2).length) {
