+++ packages/daf-cli/src/setup.ts
@@ -1,45 +1,12 @@
 import 'cross-fetch/polyfill'
 import yaml from 'yaml'
-import program from 'commander'
 import { IDataStore, IIdentityManager, IMessageHandler, IKeyManager, IResolver, TAgent } from 'daf-core'
 import { ICredentialIssuer } from 'daf-w3c'
 import { ISelectiveDisclosure } from 'daf-selective-disclosure'
 import { IDIDComm } from 'daf-did-comm'
 import { IDataStoreORM } from 'daf-typeorm'
 const fs = require('fs')
-const { dirname, resolve } = require('path')
 import { createAgentFromConfig } from './lib/agentCreator'
-
-const defaultConfig = './agent.yml'
-program.option('--config <path>', 'Configuration file', defaultConfig)
-
-program
-  .command('create-config')
-  .description('Create default agent config')
-  .option('--filename <string>', 'Config file name', './agent.yml')
-  .option('--template <string>', 'Use template (default,client)', 'default')
-
-  .action(async (options) => {
-    const { filename, template } = options
-
-    const templateFile = __dirname + '/../default/' + template + '.yml'
-    if (!fs.existsSync(templateFile)) {
-      console.log('Template not available: ' + template)
-      process.exit(1)
-    }
-
-    if (!fs.existsSync(dirname(filename))) {
-      fs.mkdirSync(dirname(filename))
-    }
-
-    if (!fs.existsSync(filename)) {
-      console.log('Creating: ' + filename)
-      const contents = fs.readFileSync(templateFile)
-      fs.writeFileSync(filename, contents)
-    } else {
-      console.log('File already exists: ' + filename)
-    }
-  })
 
 export const getConfig = (fileName: string): any => {
   if (!fs.existsSync(fileName)) {
